added cancel while editing latest mood

+++ src/redux/action.js
@@ -96,15 +96,58 @@ export const changeEntry = (idObj) => {
     }
 }
 
-export const deleteMood = (idObj) => {
-    console.log("in action for deleting mood", idObj)
+// export const deleteMood = (idObj) => {
+//     console.log("in action for deleting mood", idObj)
    
-    return function(dispatch){
-        fetch(`http://localhost:3000/api/v1/logs/${idObj}`,{
-            method: "DELETE"
-        })
-        .then(resp =>resp.json())
-        .then(data => dispatch({type: "remove mood", payload: data})) 
+//     return function(dispatch){
+//         fetch(`http://localhost:3000/api/v1/logs/${idObj}`,{
+//             method: "DELETE"
+//         })
+//         .then(resp =>resp.json())
+//         .then(data => dispatch({type: "remove mood", payload: data})) 
+//     }
+// }
+
+
+export const changeMood = (idObj, state) => {
+    console.log("in change mood action", state)
+
+    if (state.delete){
+        console.log("in action for deleting mood", idObj)
+       
+        return function(dispatch){
+            fetch(`http://localhost:3000/api/v1/logs/${idObj}`,{
+                method: "DELETE"
+            })
+            .then(resp =>resp.json())
+            .then(data => dispatch({type: "remove mood", payload: data})) 
+        }
+    } else if (state.update){
+        const obj = moodLogConverter(state)
+        console.log("updating in action", idObj)
+        return {type: "legss"}
+        // return function(dispatch){
+        //     fetch(`http://localhost:3000/api/v1/logs/${idObj}`,{
+        //         method: "PATCH",
+        //         headers: {
+        //             "content-type":"application/json",
+        //             Accept: "application/json"
+        //         },
+        //         body: JSON.stringify({
+        //             log:
+        //                 {user_id: 1,
+        //                 mood_id: obj.mood,
+        //                 tag1: obj.tag1,
+        //                 tag2: obj.tag2,
+        //                 tag3: obj.tag3,
+        //                 tag4: obj.tag4,
+        //                 tag5: obj.tag5,
+        //             }
+        //         })
+        //     })
+        //     .then(resp =>resp.json())
+        //     .then(data => dispatch({type: "add mood log", payload: data})) 
+        // }
     }
 }
